@@ -69,9 +69,9 @@
     "babel-preset-stage-2": "^6.3.13",
     "chai": "^3.4.1",
     "enzyme": "^2.0.0-rc1",
-    "eslint": "^1.10.3",
+    "eslint": "^2.2.0",
     "eslint-plugin-meteor": "^1.7.0",
-    "eslint-plugin-react": "^3.15.0",
+    "eslint-plugin-react": "^4.1.0",
     "mocha": "^2.3.4",
     "react-addons-test-utils": "^0.14.6",
     "redux-devtools": "^3.0.1",
