@@ -1,6 +1,7 @@
 import React from 'react';
 import {Main, SectionList} from '/client/modules/ui-components-v2';
-import SuitabilityMap from './components/suitability-maps.jsx';
+import InteractiveMaps from './components/interactive-maps.jsx';
+import SuitabilityMaps from './components/suitability-maps.jsx';
 import {coreLib, CoreNavList, CoreBanner} from '/client/modules/core-v2';
 
 export default (injectDeps, context) => {
@@ -18,12 +19,65 @@ export default (injectDeps, context) => {
           displayName: 'SectionList',
           spacing: false,
           sections: [
-            // () => (React.createElement(CoreBanner, {
-            //   displayName: 'CoreBanner',
-            //   path: this.name
-            // })),
-            () => (React.createElement(SuitabilityMap, {
-              displayName: 'SuitabilityMap'
+            () => (React.createElement(InteractiveMaps, {
+              displayName: 'InteractiveMaps'
+            }))
+          ]
+        }))
+      }));
+    }
+  });
+
+  FlowRouter.route('/suitability-maps', {
+    name: 'suitability-maps.root',
+    triggersEnter: [triggerAuthenticate],
+    action() {
+      mount(MainCtx, Object.assign({}, layout, {
+        content: () => (React.createElement(SectionList, {
+          displayName: 'SectionList',
+          spacing: false,
+          sections: [
+            () => (React.createElement(SuitabilityMaps, {
+              displayName: 'SuitabilityMaps'
+            }))
+          ]
+        }))
+      }));
+    }
+  });
+
+  FlowRouter.route('/suitability-maps/:crop', {
+    name: 'suitability-maps.root',
+    triggersEnter: [triggerAuthenticate],
+    action() {
+      mount(MainCtx, Object.assign({}, layout, {
+        content: () => (React.createElement(SectionList, {
+          displayName: 'SectionList',
+          spacing: false,
+          sections: [
+            () => (React.createElement(SuitabilityMaps, {
+              displayName: 'SuitabilityMaps',
+              crop: FlowRouter.getParam('crop')
+            }))
+          ]
+        }))
+      }));
+    }
+  });
+
+  FlowRouter.route('/suitability-maps/:crop/:suitability', {
+    name: 'suitability-maps.root',
+    triggersEnter: [triggerAuthenticate],
+    action() {
+      mount(MainCtx, Object.assign({}, layout, {
+        content: () => (React.createElement(SectionList, {
+          displayName: 'SectionList',
+          spacing: false,
+          sections: [
+            () => (React.createElement(SuitabilityMaps, {
+              displayName: 'SuitabilityMaps',
+              crop: FlowRouter.getParam('crop'),
+              suitable: FlowRouter.getParam('suitability')
             }))
           ]
         }))
