@@ -1,4 +1,7 @@
+import React from 'react';
+import {MdlLanding} from '/client/modules/mdl-core-design';
+
 export default (injectDeps, context) => {
-  const {} = context;
-  injectDeps();
+  const {FlowRouter, mount} = context;
+  const MdlLandingCtx = injectDeps(MdlLanding);
 };
