@@ -1,6 +1,7 @@
+// Will be part of Version 1.2
 import {Mongo} from 'meteor/mongo';
 
-const Admin = new Mongo.Collection('admin');
+const Admin = new Mongo.Collection('sarai-admin');
 
 Admin.deny({
   insert: () => true,
