@@ -1,14 +1,14 @@
 // will be part of version 1.2
 import {Mongo} from 'meteor/mongo';
 
-const NavMenu = new Mongo.Collection('nav-menu');
+const Banner = new Mongo.Collection('sarai-banner');
 
-NavMenu.deny({
+Banner.deny({
   insert: () => true,
   update: () => true,
   remove: () => true
 });
 
-NavMenu.methods = {};
+Banner.methods = {};
 
-export default NavMenu;
+export default Banner;
