@@ -1,21 +1,47 @@
-import _LandingData from './landing-data';
-import _KnowledgeHub from './knowledge-hub';
-import _Admin from './admin';
-import _NavMenu from './nav-menu';
-import _Cms from './cms';
-import _StationList from './station-list';
-import _StationData from './station-data';
-import _WeatherStations from './weather-stations';
-import _SimulatedYield from './simulated-yield';
-import _ChartLabels from './chart-labels';
+// <<<<<<< HEAD
+// import _LandingData from './landing-data';
+// import _KnowledgeHub from './knowledge-hub';
+// import _Admin from './admin';
+// import _NavMenu from './nav-menu';
+// import _Cms from './cms';
+// import _StationList from './station-list';
+// import _StationData from './station-data';
+// import _WeatherStations from './weather-stations';
+// import _SimulatedYield from './simulated-yield';
+// import _ChartLabels from './chart-labels';
 
-export const Admin = _Admin;
-export const LandingData = _LandingData;
-export const KnowledgeHub = _KnowledgeHub;
-export const NavMenu = _NavMenu;
-export const Cms = _Cms;
-export const StationList = _StationList;
-export const StationData = _StationData;
-export const WeatherStations = _WeatherStations;
-export const SimulatedYield = _SimulatedYield;
-export const ChartLabels = _ChartLabels;
+// export const Admin = _Admin;
+// export const LandingData = _LandingData;
+// export const KnowledgeHub = _KnowledgeHub;
+// export const NavMenu = _NavMenu;
+// export const Cms = _Cms;
+// export const StationList = _StationList;
+// export const StationData = _StationData;
+// export const WeatherStations = _WeatherStations;
+// export const SimulatedYield = _SimulatedYield;
+// export const ChartLabels = _ChartLabels;
+// =======
+import LandingData from './landing-data';
+import Admin from './admin';
+import NavMenu from './nav-menu';
+import Pages from './pages';
+import Banner from './banner';
+import StationList from './station-list';
+import StationData from './station-data';
+import SimulatedYield from './simulated-yield';
+import ChartLabels from './chart-labels';
+
+export {
+  Admin,
+  NavMenu,
+  Pages,
+  Banner,
+  StationList,
+  StationData,
+  SimulatedYield,
+  ChartLabels,
+
+  // Deprecated
+  LandingData
+};
+// >>>>>>> develop
