@@ -1,6 +1,7 @@
+// will be part of version 1.2
 import {Mongo} from 'meteor/mongo';
 
-const PagesData = new Mongo.Collection('pages-write-up');
+const PagesData = new Mongo.Collection('sarai-pages');
 
 PagesData.deny({
   insert: () => true,
